--- composeApp/src/androidMain/kotlin/com/leoapps/eggy/Platform.android.kt
@@ -1,9 +1,0 @@
-package com.leoapps.eggy
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
