--- composeApp/src/commonMain/kotlin/com/leoapps/eggy/Greeting.kt
@@ -1,9 +1,0 @@
-package com.leoapps.eggy
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
