--- composeApp/src/commonMain/kotlin/com/leoapps/eggy/Platform.kt
@@ -1,7 +1,0 @@
-package com.leoapps.eggy
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
