+++ composeApp/src/commonMain/kotlin/com/leoapps/eggy/common/permissions/model/PermissionStatus.kt
@@ -1,4 +1,4 @@
-package com.leoapps.eggy.base.permissions.model
+package com.leoapps.eggy.common.permissions.model
 
 enum class PermissionStatus {
     GRANTED,
