+++ composeApp/src/commonMain/kotlin/com/leoapps/eggy/common/vibration/presentation/LocalVibrationManager.kt
@@ -1,7 +1,7 @@
 package com.leoapps.vibration.presentation
 
 import androidx.compose.runtime.compositionLocalOf
-import com.leoapps.eggy.vibration.domain.VibrationManager
+import com.leoapps.eggy.common.vibration.domain.VibrationManager
 
 val LocalVibrationManager = compositionLocalOf<VibrationManager> {
     error("No WaterAppVibrator provided")
