--- composeApp/src/iosMain/kotlin/com/leoapps/eggy/Platform.ios.kt
@@ -1,9 +1,0 @@
-package com.leoapps.eggy
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
