+++ composeApp/src/iosMain/kotlin/com/leoapps/eggy/common/vibrator/VibrationManagerIosImpl.kt
@@ -1,6 +1,6 @@
-package com.leoapps.eggy.vibrator
+package com.leoapps.eggy.common.vibrator
 
-import com.leoapps.eggy.vibration.domain.VibrationManager
+import com.leoapps.eggy.common.vibration.domain.VibrationManager
 import platform.UIKit.UIImpactFeedbackGenerator
 import platform.UIKit.UIImpactFeedbackStyle
 import platform.UIKit.UINotificationFeedbackGenerator
